Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> At first glance it looks like this test suite could leverage Robot framework - I can see a lot of bash scripts performing requests and asserting the response - I guess it is a perfect usecase for Robot, no?

These suites are used and maintained as such by Bell. They used them in their env.
They are sharing them with the community in order to improve the test coverage of CDS.

So I do not see the a big interest to translate them into robot (or in postman or pythonsdk or any other framework)..
the most important is the content of the tests (the url..) not necessarily the framework used to launch them.
Using robot would bring additional built-in features and could be seen as a consistency improvement with some our existing tests (note most of the robot healtcheck tests are useless because already covered by kubernetes native readiness and liveness tests)
So if people find time to do it why not but for me it is not a top priority.

Patch-set: 1
Reviewer: Gerrit User 1215 <1215@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>