Update patch set 6
Patch Set 6:

> Patch Set 6: Code-Review-1
> 
> 1) Look at Comment Dan posted:
> 
> Just to be clear, am I correct in understanding these bash scripts would be run OUTSIDE of any docker container. My main concern is whether this would require us to install bash (licensed under LGPL v3) as part of our ONAP dockers.
> 
> 2) Is it possible to split this review in multiple chunks ?

it is a baseline of the CDS tests used by Bell and shared with the community
for the moment it is just a set of bash/json/yml/kt/python  scripts to be run on a CDS system (even independently from ONAP)

It is not dockerized, not packaged, not integrated in any test in CI
There is no plan for distribution for the moment, at least nothing for G release.

We need to discuss with people who submitted the patch how we can use it the most efficiently
shall write robot tests, update the cds support in pythonsdk? just run the different sh script?
shall it be done at the CSIT level (CDS standalone) or in the integration chains (ONAP context)

meanwhile I prefer to get them versioned in a repo rather than in a zip file or in the wiki.

@Santos/Sébastien, could we plan a demo during weekly integration meeting to see how you are using these tests within Bell context and see how we can leverage them for the community in H release?

Patch-set: 6
Reviewer: Gerrit User 1215 <1215@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>